ctservice.js: Fetch user IDs for group members only if required
If we want to export only users from certain groups specified in the
config, we need to query the global group members endpoint. But if
we want to fetch all users that are visible to us, we don't need this
step.

This will allow us to run without the "administer persons" permission
that is required for the global group members endpoint.