Revert to use global group members endpoint and filter departments
Previous work to avoid the global group members endpoint caused high
load on the API server as we needed an additional request per group.

Instead we now revert to the old behavior which requires to give the
API user "administer persons" permissions and therefore see all persons.

In order to still limit the exported persons to a certain department,
let's introduce a department filter for the persons.